# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/invoices/bulk": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Create new invoices in bulk mode",
                "operationId": "e302bbc8a2e0b114012e06c1977c08ff",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "anyOf": [
                                        {
                                            "required": [
                                                "customerId",
                                                "amount",
                                                "billed_date",
                                                "status"
                                            ],
                                            "properties": {
                                                "customerId": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": 1235
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "B"
                                                },
                                                "billedDate": {
                                                    "type": "dateTime",
                                                    "example": "2017-09-14 22:55:56"
                                                },
                                                "paidDate": {
                                                    "type": "dateTime",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "required": [
                                                "customerId",
                                                "amount",
                                                "billed_date",
                                                "status"
                                            ],
                                            "properties": {
                                                "customerId": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": 4567
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "P"
                                                },
                                                "billedDate": {
                                                    "type": "dateTime",
                                                    "example": "2017-09-14 22:55:56"
                                                },
                                                "paidDate": {
                                                    "type": "dateTime",
                                                    "example": "2017-09-14 22:55:56"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Unauthenticated."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token starting from Bearer e.g. `Bearer 12345678`",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}